refactor: header props 관련 로직 Customhook으로 분리
